# New level-based L0->base_level compaction picking algorithm