Previously, we picked largest L0 file and all overlapping ones as inputs for L0->base compaction. It didn't seem to have any benefits and L0->base couldn't run concurrently with L0->L0.

This diff changes L0->base picker to start with the oldest L0 file, then expand towards newer L0 files until reaching a file being compacted. Now, L0->L0 can run concurrently since that one's picker starts with the newest file and expands towards older L0 files until reaching a file being compacted. This prevents overlap/conflict.

Test Plan:
- unit tests; updated the intra-L0 test to verify simultaneous L0->base and L0->L0.
- fillseq benchmark (to exercise trivial move from L0)
```
command: TEST_TMPDIR=/data/compaction_bench/  ./db_bench -benchmarks=fillseq -options_file=/home/andrewkr/myrocks-options-modified -statistics -num=100000000
before: fillseq      :       2.041 micros/op 489997 ops/sec;   54.2 MB/s
after: fillseq      :       2.026 micros/op 493651 ops/sec;   54.6 MB/s
```