@siying 2017-04-06
This is not a refactoring, but a behavior change, right?

From: Andrew Kryczka [mailto:notifications@github.com]
Sent: Thursday, April 6, 2017 2:59 PM
To: facebook/rocksdb <rocksdb@noreply.github.com>
Cc: Subscribed <subscribed@noreply.github.com>
Subject: [facebook/rocksdb] Refactor L0->base_level compaction (#2103)


Previously, we picked largest L0 file and all overlapping ones as inputs to L0->base. It didn't seem to have any benefits and L0->base couldn't run concurrently with L0->L0.

This diff changes L0->base picker to start with the oldest L0 file, then expand towards newer L0 files until reaching a file being compacted. Now, L0->L0 can run concurrently since that one's picker starts with the newest file and expands towards older L0 files. This prevents overlap/conflict.

But I haven't removed the restriction (level0_compactions_in_progress_) yet. That'll be another diff.

Test Plan: unit tests. In the next diff I'll test simultaneous L0->base and L0->L0.

________________________________
You can view, comment on, or merge this pull request online at:

  https://github.com/facebook/rocksdb/pull/2103

Commit Summary

  *   Refactor L0->base_level compaction

File Changes

  *   M db/compaction_picker.cc<https://github.com/facebook/rocksdb/pull/2103/files#diff-0> (127)
  *   M db/compaction_picker.h<https://github.com/facebook/rocksdb/pull/2103/files#diff-1> (11)
  *   M db/db_compaction_test.cc<https://github.com/facebook/rocksdb/pull/2103/files#diff-2> (91)

Patch Links:

  *   https://github.com/facebook/rocksdb/pull/2103.patch
  *   https://github.com/facebook/rocksdb/pull/2103.diff

—
You are receiving this because you are subscribed to this thread.
Reply to this email directly, view it on GitHub<https://github.com/facebook/rocksdb/pull/2103>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AFFmJyf_TomvV03z9-h_kPzyuw3oj-Acks5rtWBAgaJpZM4M2OGP>.