@ajkr 2017-04-06
sure I updated the title. The goal is more to simplify the algorithm (although as usual the resulting code looks more complicated), not really any perf improvement or user-visible behavior change